修复 Dart 移植两处编译错误
- natural_language_entry_parser：jiao/Decimal(10) 返回 Rational，补 .toDecimal()
- budget_engine_test：去掉数字字面量下划线分隔符（当前 Dart 版本不支持）

+++ android-app/lib/core/ai/natural_language_entry_parser.dart
@@ -69,7 +69,7 @@ class NaturalLanguageEntryParser {
       final yuan = Decimal.tryParse(kuaiMatch.group(1)!);
       final jiao = Decimal.tryParse(kuaiMatch.group(2)!);
       if (yuan != null && jiao != null) {
-        return yuan + jiao / Decimal.fromInt(10);
+        return yuan + (jiao / Decimal.fromInt(10)).toDecimal();
       }
     }
